pyproject.toml
[project]
name = "video-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "torchvision",
    "torchaudio",
    "opencv-python>=4.13.0.92",
    "python-dotenv>=1.2.1",
    "transformers>=5.1.0",
    "accelerate>=1.12.0",
    "pillow>=12.1.0",
    "chromadb>=1.4.1",
    "ollama>=0.6.1",
    "langchain>=1.2.9",
    "langchain-groq>=1.1.2",
    "moviepy>=1.0.3",
    "sentence-transformers>=5.2.2",
    "tqdm>=4.67.3",
    "scenedetect>=0.6.7.1",
    "scikit-image>=0.26.0",
    "ipykernel>=7.2.0",
    "easyocr>=1.7.2",
    "openai-whisper>=20250625",
    "faster-whisper>=1.2.1",
    "langchain-text-splitters>=1.1.0",
    "fastapi>=0.128.7",
    "pydantic>=2.12.5",
    "supermemory>=3.24.0",
    "platformdirs>=4.5.1",
    "typer>=0.21.1",
]   

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }


[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
video-rag = "videorag.cli.main:app"

